Named Colors

The names of available colors are stored in alphabetical order in the dictionary Colors.color_names:

color_names = Dict(
    "aliceblue"            => (240, 248, 255),
    "antiquewhite"         => (250, 235, 215),
    "antiquewhite1"        => (255, 239, 219),
    ...

Info

Colors.jl supports the CSS/SVG named colors and the X11 named colors. The CSS/SVG named colors come from the 16 colors defined in HTML3.2 and the X11 named colors. There are some unnatural definitions due to the different origins. For example, "LightGray" is lighter than "Gray", but "DarkGray" is also lighter than "Gray".

Julia logo colors

Colors.JULIA_LOGO_COLORS is a NamedTuple containing the julia logo colors defined in the julia-logo-graphics.

logocolors = Colors.JULIA_LOGO_COLORS
[logocolors.blue, logocolors.red, logocolors.green, logocolors.purple]

Colors.JULIA_LOGO_COLORS — Constant

Colors used in the Julia logo as a NamedTuple.

The keys are approximate descriptions of the hue and do not include black.

Not exported, use as JULIA_LOGO_COLORS.red etc.